refactor: use transaction identifiers for transaction method
