+++ app/DTO/Inertia/TransactionDetails.php
@@ -32,8 +32,6 @@ class TransactionDetails extends Data
         #[LiteralTypeScriptType('{spender: IWalletReference; amount: string | null; isUnlimited: boolean; isRevoke: boolean} | null')]
         public ?array $tokenApproval,
         public ?Token $token,
-        #[LiteralTypeScriptType('{formatted: string | null; utf8: string | null; raw: string | null} | null')]
-        public ?array $payload,
         #[LiteralTypeScriptType('{recipient: IWalletReference; amount: string}[]')]
         public array $batchTokenTransfers,
     ) {
@@ -96,7 +94,6 @@ class TransactionDetails extends Data
             tokenTransfer: self::tokenTransferDetails($viewModel),
             tokenApproval: self::tokenApprovalDetails($viewModel),
             token: $token,
-            payload: self::payloadDetails($viewModel),
             batchTokenTransfers: $batchTokenTransfers,
         );
     }
@@ -131,39 +128,5 @@ class TransactionDetails extends Data
             'recipient' => $recipientWalletData,
             'amount'    => $amount,
         ];
-    }
-
-    /**
-     * @return array{formatted: ?string, utf8: ?string, raw: ?string}|null
-     */
-    private static function payloadDetails(TransactionViewModel $transaction): ?array
-    {
-        if (! $transaction->hasPayload()) {
-            return null;
-        }
-
-        return [
-            // Ensure invalid UTF-8 does not break JSON serialization.
-            'formatted' => self::safeUtf8($transaction->formattedPayload() ?? ''),
-            'utf8'      => self::safeUtf8($transaction->utf8Payload() ?? ''),
-            'raw'       => self::safeUtf8($transaction->rawPayload() ?? ''),
-        ];
-    }
-
-    private static function safeUtf8(string $value): string
-    {
-        // Skip normalization when the payload is already valid UTF-8.
-        if (preg_match('//u', $value) === 1) {
-            return $value;
-        }
-
-        // Invalid UTF-8 in payloads breaks JSON encoding and causes Inertia JSON.parse errors;
-        // normalize to the replacement character to keep the response valid.
-        $previousSubstitute = mb_substitute_character();
-        mb_substitute_character(0xFFFD);
-        $converted = mb_convert_encoding($value, 'UTF-8', 'UTF-8');
-        mb_substitute_character($previousSubstitute);
-
-        return $converted;
     }
 }
